Revert "Se crea conexión de front con back para testear endoints"
