--- electron-app/dist-electron/db/database.js
@@ -1,75 +1,0 @@
-"use strict";
-var __createBinding = (this && this.__createBinding) || (Object.create ? (function(o, m, k, k2) {
-    if (k2 === undefined) k2 = k;
-    var desc = Object.getOwnPropertyDescriptor(m, k);
-    if (!desc || ("get" in desc ? !m.__esModule : desc.writable || desc.configurable)) {
-      desc = { enumerable: true, get: function() { return m[k]; } };
-    }
-    Object.defineProperty(o, k2, desc);
-}) : (function(o, m, k, k2) {
-    if (k2 === undefined) k2 = k;
-    o[k2] = m[k];
-}));
-var __setModuleDefault = (this && this.__setModuleDefault) || (Object.create ? (function(o, v) {
-    Object.defineProperty(o, "default", { enumerable: true, value: v });
-}) : function(o, v) {
-    o["default"] = v;
-});
-var __importStar = (this && this.__importStar) || (function () {
-    var ownKeys = function(o) {
-        ownKeys = Object.getOwnPropertyNames || function (o) {
-            var ar = [];
-            for (var k in o) if (Object.prototype.hasOwnProperty.call(o, k)) ar[ar.length] = k;
-            return ar;
-        };
-        return ownKeys(o);
-    };
-    return function (mod) {
-        if (mod && mod.__esModule) return mod;
-        var result = {};
-        if (mod != null) for (var k = ownKeys(mod), i = 0; i < k.length; i++) if (k[i] !== "default") __createBinding(result, mod, k[i]);
-        __setModuleDefault(result, mod);
-        return result;
-    };
-})();
-var __importDefault = (this && this.__importDefault) || function (mod) {
-    return (mod && mod.__esModule) ? mod : { "default": mod };
-};
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.initializeDatabase = initializeDatabase;
-exports.getDatabase = getDatabase;
-const sqlite3_1 = __importDefault(require("sqlite3"));
-const sqlite_1 = require("sqlite");
-const path = __importStar(require("path"));
-sqlite3_1.default.verbose();
-let db = null;
-async function initializeDatabase() {
-    if (!db) {
-        const dbPath = path.join(process.cwd(), 'electron', 'db', 'mydb.sqlite');
-        console.log('Intentando abrir base de datos en:', dbPath);
-        db = await (0, sqlite_1.open)({
-            filename: dbPath,
-            driver: sqlite3_1.default.Database
-        });
-        await db.exec(`
-      CREATE TABLE IF NOT EXISTS usuario (
-        id INTEGER PRIMARY KEY AUTOINCREMENT,
-        nombre TEXT NOT NULL,
-        usuario TEXT NOT NULL,
-        correo TEXT,
-        telefono TEXT NOT NULL,
-        direccion TEXT NOT NULL,
-        password TEXT NOT NULL,
-        rol TEXT CHECK (rol IN ('Administrador', 'ServicioSocial', 'Albergue')) NOT NULL
-      )
-    `);
-    }
-    return db;
-}
-async function getDatabase() {
-    if (!db) {
-        await initializeDatabase();
-    }
-    return db;
-}
-//# sourceMappingURL=database.js.map
