--- electron-app/dist-electron/domains/users/users.model.js
@@ -1,3 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-//# sourceMappingURL=users.model.js.map
