--- electron-app/dist-electron/main.cjs
@@ -1,97 +1,0 @@
-"use strict";
-var __createBinding = (this && this.__createBinding) || (Object.create ? (function(o, m, k, k2) {
-    if (k2 === undefined) k2 = k;
-    var desc = Object.getOwnPropertyDescriptor(m, k);
-    if (!desc || ("get" in desc ? !m.__esModule : desc.writable || desc.configurable)) {
-      desc = { enumerable: true, get: function() { return m[k]; } };
-    }
-    Object.defineProperty(o, k2, desc);
-}) : (function(o, m, k, k2) {
-    if (k2 === undefined) k2 = k;
-    o[k2] = m[k];
-}));
-var __setModuleDefault = (this && this.__setModuleDefault) || (Object.create ? (function(o, v) {
-    Object.defineProperty(o, "default", { enumerable: true, value: v });
-}) : function(o, v) {
-    o["default"] = v;
-});
-var __importStar = (this && this.__importStar) || (function () {
-    var ownKeys = function(o) {
-        ownKeys = Object.getOwnPropertyNames || function (o) {
-            var ar = [];
-            for (var k in o) if (Object.prototype.hasOwnProperty.call(o, k)) ar[ar.length] = k;
-            return ar;
-        };
-        return ownKeys(o);
-    };
-    return function (mod) {
-        if (mod && mod.__esModule) return mod;
-        var result = {};
-        if (mod != null) for (var k = ownKeys(mod), i = 0; i < k.length; i++) if (k[i] !== "default") __createBinding(result, mod, k[i]);
-        __setModuleDefault(result, mod);
-        return result;
-    };
-})();
-Object.defineProperty(exports, "__esModule", { value: true });
-const electron_1 = require("electron");
-const path = __importStar(require("path"));
-const index_js_1 = require("./domains/index.js");
-const database_js_1 = require("./db/database.js");
-let mainWindow = null;
-async function createWindow() {
-    try {
-        // Inicializar la base de datos primero
-        await (0, database_js_1.initializeDatabase)();
-        console.log('Base de datos inicializada correctamente');
-        mainWindow = new electron_1.BrowserWindow({
-            width: 800,
-            height: 600,
-            webPreferences: {
-                preload: path.join(__dirname, 'preload.cjs'),
-                contextIsolation: true,
-                nodeIntegration: false
-            }
-        });
-        mainWindow.loadURL('http://localhost:5173');
-        const template = [
-            {
-                label: 'Archivo',
-                submenu: [{ role: 'quit', label: 'Salir' }]
-            },
-            {
-                label: 'Editar',
-                submenu: [
-                    { role: 'undo', label: 'Deshacer' },
-                    { role: 'redo', label: 'Rehacer' },
-                    { type: 'separator' },
-                    { role: 'cut', label: 'Cortar' },
-                    { role: 'copy', label: 'Copiar' },
-                    { role: 'paste', label: 'Pegar' }
-                ]
-            },
-            {
-                label: 'Ver',
-                submenu: [
-                    { role: 'reload', label: 'Recargar' },
-                    { role: 'toggleDevTools', label: 'Herramientas de desarrollo' },
-                    { type: 'separator' },
-                    { role: 'resetZoom', label: 'Restablecer zoom' },
-                    { role: 'zoomIn', label: 'Acercar' },
-                    { role: 'zoomOut', label: 'Alejar' },
-                    { type: 'separator' },
-                    { role: 'togglefullscreen', label: 'Pantalla completa' }
-                ]
-            }
-        ];
-        const menu = electron_1.Menu.buildFromTemplate(template);
-        electron_1.Menu.setApplicationMenu(menu);
-        // ⬇️ Inicializa el backend (tu "Nest-like")
-        await (0, index_js_1.bootstrapBackend)(electron_1.ipcMain);
-    }
-    catch (error) {
-        console.error('Error al inicializar la aplicación:', error);
-        electron_1.app.quit();
-    }
-}
-electron_1.app.whenReady().then(createWindow);
-//# sourceMappingURL=main.js.map
