+++ electron-app/dist-electron/main.js
@@ -1,0 +1,51 @@
+import { app, BrowserWindow, Menu } from 'electron';
+import { fileURLToPath } from 'url';
+import { dirname } from 'path';
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+function createWindow() {
+    const win = new BrowserWindow({
+        width: 800,
+        height: 600,
+        webPreferences: {
+            preload: __dirname + '/preload.js', // más adelante lo cambiamos a .ts compilado
+            contextIsolation: true,
+            nodeIntegration: false
+        }
+    });
+    win.loadURL('http://localhost:5173');
+    const template = [
+        {
+            label: 'Archivo',
+            submenu: [{ role: 'quit', label: 'Salir' }]
+        },
+        {
+            label: 'Editar',
+            submenu: [
+                { role: 'undo', label: 'Deshacer' },
+                { role: 'redo', label: 'Rehacer' },
+                { type: 'separator' },
+                { role: 'cut', label: 'Cortar' },
+                { role: 'copy', label: 'Copiar' },
+                { role: 'paste', label: 'Pegar' }
+            ]
+        },
+        {
+            label: 'Ver',
+            submenu: [
+                { role: 'reload', label: 'Recargar' },
+                { role: 'toggleDevTools', label: 'Herramientas de desarrollo' },
+                { type: 'separator' },
+                { role: 'resetZoom', label: 'Restablecer zoom' },
+                { role: 'zoomIn', label: 'Acercar' },
+                { role: 'zoomOut', label: 'Alejar' },
+                { type: 'separator' },
+                { role: 'togglefullscreen', label: 'Pantalla completa' }
+            ]
+        }
+    ];
+    const menu = Menu.buildFromTemplate(template);
+    Menu.setApplicationMenu(menu);
+}
+app.whenReady().then(createWindow);
+//# sourceMappingURL=main.js.map
