+++ electron-app/dist-electron/preload.js
@@ -1,0 +1,5 @@
+import { contextBridge } from 'electron';
+contextBridge.exposeInMainWorld('api', {
+    ping: () => 'pong'
+});
+//# sourceMappingURL=preload.js.map
