--- electron-app/electron/domains/users/users.model.ts
@@ -1,10 +1,0 @@
-export interface Usuario {
-  id: number;
-  nombre: string;
-  usuario: string;
-  correo?: string;
-  telefono: string;
-  direccion: string;
-  password: string;
-  rol: 'Administrador' | 'ServicioSocial' | 'Albergue';
-}
