+++ electron-app/electron/main.ts
@@ -1,68 +1,56 @@
-import { app, BrowserWindow, Menu, ipcMain } from 'electron'
-import * as path from 'path';
-import { bootstrapBackend } from './domains/index.js';
-import { initializeDatabase } from './db/database.js';
+import { app, BrowserWindow, Menu } from 'electron'
+import { fileURLToPath } from 'url';
+import { dirname } from 'path';
 
-let mainWindow: BrowserWindow | null = null;
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
 
-async function createWindow() {
-  try {
-    // Inicializar la base de datos primero
-    await initializeDatabase();
-    console.log('Base de datos inicializada correctamente');
+function createWindow() {
+  const win = new BrowserWindow({
+    width: 800,
+    height: 600,
+    webPreferences: {
+      preload: __dirname + '/preload.js', // más adelante lo cambiamos a .ts compilado
+      contextIsolation: true,
+      nodeIntegration: false
+    }
+  })
 
-    mainWindow = new BrowserWindow({
-      width: 800,
-      height: 600,
-      webPreferences: {
-        preload: path.join(__dirname, 'preload.cjs'),
-        contextIsolation: true,
-        nodeIntegration: false
-      }
-    })
+  win.loadURL('http://localhost:5173')
 
-    mainWindow.loadURL('http://localhost:5173')
+  const template: Electron.MenuItemConstructorOptions[] = [
+    {
+      label: 'Archivo',
+      submenu: [{ role: 'quit', label: 'Salir' }]
+    },
+    {
+      label: 'Editar',
+      submenu: [
+        { role: 'undo', label: 'Deshacer' },
+        { role: 'redo', label: 'Rehacer' },
+        { type: 'separator' },
+        { role: 'cut', label: 'Cortar' },
+        { role: 'copy', label: 'Copiar' },
+        { role: 'paste', label: 'Pegar' }
+      ]
+    },
+    {
+      label: 'Ver',
+      submenu: [
+        { role: 'reload', label: 'Recargar' },
+        { role: 'toggleDevTools', label: 'Herramientas de desarrollo' },
+        { type: 'separator' },
+        { role: 'resetZoom', label: 'Restablecer zoom' },
+        { role: 'zoomIn', label: 'Acercar' },
+        { role: 'zoomOut', label: 'Alejar' },
+        { type: 'separator' },
+        { role: 'togglefullscreen', label: 'Pantalla completa' }
+      ]
+    }
+  ]
 
-    const template: Electron.MenuItemConstructorOptions[] = [
-      {
-        label: 'Archivo',
-        submenu: [{ role: 'quit', label: 'Salir' }]
-      },
-      {
-        label: 'Editar',
-        submenu: [
-          { role: 'undo', label: 'Deshacer' },
-          { role: 'redo', label: 'Rehacer' },
-          { type: 'separator' },
-          { role: 'cut', label: 'Cortar' },
-          { role: 'copy', label: 'Copiar' },
-          { role: 'paste', label: 'Pegar' }
-        ]
-      },
-      {
-        label: 'Ver',
-        submenu: [
-          { role: 'reload', label: 'Recargar' },
-          { role: 'toggleDevTools', label: 'Herramientas de desarrollo' },
-          { type: 'separator' },
-          { role: 'resetZoom', label: 'Restablecer zoom' },
-          { role: 'zoomIn', label: 'Acercar' },
-          { role: 'zoomOut', label: 'Alejar' },
-          { type: 'separator' },
-          { role: 'togglefullscreen', label: 'Pantalla completa' }
-        ]
-      }
-    ]
-
-    const menu = Menu.buildFromTemplate(template)
-    Menu.setApplicationMenu(menu)
-
-    // ⬇️ Inicializa el backend (tu "Nest-like")
-    await bootstrapBackend(ipcMain);
-  } catch (error) {
-    console.error('Error al inicializar la aplicación:', error);
-    app.quit();
-  }
+  const menu = Menu.buildFromTemplate(template)
+  Menu.setApplicationMenu(menu)
 }
 
 app.whenReady().then(createWindow)
