+++ electron-app/electron/preload.ts
@@ -1,10 +1,5 @@
-import { contextBridge, ipcRenderer } from 'electron';
+import { contextBridge } from 'electron'
 
 contextBridge.exposeInMainWorld('api', {
-  ping: () => ipcRenderer.invoke('ping'),
-
-  // Ejemplo: Llamada para obtener todos los usuarios
-  getUsuarios: () => ipcRenderer.invoke('usuarios:getAll'),
-
-  // Puedes seguir agregando más métodos aquí
-});
+  ping: () => 'pong'
+})
