--- electron-app/src/components/Footer.tsx
@@ -1,11 +1,0 @@
-const Footer = () => {
-    return (
-        <footer className="bg-quinto font-bold py-2 mt-auto">
-            <div className="container mx-auto text-center items-center">
-                <p className="text-xs">© 2025 Sumate Al Amor. Todos los derechos reservados.</p>
-            </div>
-        </footer>
-    )
-}
-
-export default Footer
