+++ electron-app/src/index.css
@@ -14,12 +14,7 @@
   --font-nunito: 'Nunito', sans-serif;
 }
 
-h1,
-h2,
-h3,
-h4,
-h5,
-h6 {
+h1, h2, h3, h4, h5, h6 {
   font-family: var(--font-poppins);
 }
 
@@ -28,66 +23,25 @@ p {
 }
 
 /* Colores de texto */
-.primario {
-  color: var(--color-primario);
-}
-
-.secundario {
-  color: var(--color-secundario);
-}
-
-.terciario {
-  color: var(--color-terciario);
-}
-
-.cuarto {
-  color: var(--color-cuarto);
-}
-
-.quinto {
-  color: var(--color-quinto);
-}
-
-.sexto {
-  color: var(--color-sexto);
-}
-
-.septimo {
-  color: var(--color-septimo);
-}
+.primario { color: var(--color-primario); }
+.secundario { color: var(--color-secundario); }
+.terciario { color: var(--color-terciario); }
+.cuarto { color: var(--color-cuarto); }
+.quinto { color: var(--color-quinto); }
+.sexto { color: var(--color-sexto); }
+.septimo { color: var(--color-septimo); }
 
 /* Colores de fondo */
-.bg-primario {
-  background-color: var(--color-primario);
-}
-
-.bg-secundario {
-  background-color: var(--color-secundario);
-}
-
-.bg-terciario {
-  background-color: var(--color-terciario);
-}
-
-.bg-cuarto {
-  background-color: var(--color-cuarto);
-}
-
-.bg-quinto {
-  background-color: var(--color-quinto);
-}
-
-.bg-sexto {
-  background-color: var(--color-sexto);
-}
-
-.bg-septimo {
-  background-color: var(--color-septimo);
-}
+.bg-primario { background-color: var(--color-primario); }
+.bg-secundario { background-color: var(--color-secundario); }
+.bg-terciario { background-color: var(--color-terciario); }
+.bg-cuarto { background-color: var(--color-cuarto); }
+.bg-quinto { background-color: var(--color-quinto); }
+.bg-sexto { background-color: var(--color-sexto); }
+.bg-septimo { background-color: var(--color-septimo); }
 
 /* Layout para footer y navbar */
-html,
-body {
+html, body {
   height: 100%;
   margin: 0;
   font-family: var(--font-nunito);
@@ -95,79 +49,18 @@ body {
   flex-direction: column;
 }
 
-/* Hover: background-color */
-.hover\:bg-primario:hover {
-  background-color: var(--color-primario);
-}
-
-.hover\:bg-secundario:hover {
-  background-color: var(--color-secundario);
-}
-
-.hover\:bg-terciario:hover {
-  background-color: var(--color-terciario);
-}
-
-.hover\:bg-cuarto:hover {
-  background-color: var(--color-cuarto);
-}
-
-.hover\:bg-quinto:hover {
-  background-color: var(--color-quinto);
-}
-
-.hover\:bg-sexto:hover {
-  background-color: var(--color-sexto);
-}
-
-.hover\:bg-septimo:hover {
-  background-color: var(--color-septimo);
-}
-
-/* Hover: text color */
-.hover\:text-primario:hover {
-  color: var(--color-primario);
-}
-
-.hover\:text-secundario:hover {
-  color: var(--color-secundario);
-}
-
-.hover\:text-terciario:hover {
-  color: var(--color-terciario);
-}
-
-.hover\:text-cuarto:hover {
-  color: var(--color-cuarto);
-}
-
-.hover\:text-quinto:hover {
-  color: var(--color-quinto);
-}
-
-.hover\:text-sexto:hover {
-  color: var(--color-sexto);
-}
-
-.hover\:text-septimo:hover {
-  color: var(--color-septimo);
-}
-
-
 #navbar {
   position: fixed;
   top: 0;
   left: 0;
   width: 100%;
-  height: 20px;
-  /* ajusta según necesites */
+  height: 60px; /* ajusta según necesites */
   background-color: var(--color-primario);
   color: white;
   display: flex;
   align-items: center;
   padding: 0 20px;
   z-index: 1000;
-  box-sizing: border-box;
 }
 
 #footer {
@@ -175,8 +68,7 @@ body {
   bottom: 0;
   left: 0;
   width: 100%;
-  height: 50px;
-  /* ajusta según necesites */
+  height: 50px; /* ajusta según necesites */
   background-color: var(--color-secundario);
   color: white;
   display: flex;
@@ -188,9 +80,7 @@ body {
 #content {
   /* Para que el contenido no quede tapado por navbar ni footer */
   flex: 1;
-  margin-top: 60px;
-  /* igual que la altura del navbar */
-  margin-bottom: 50px;
-  /* igual que la altura del footer */
+  margin-top: 60px;  /* igual que la altura del navbar */
+  margin-bottom: 50px; /* igual que la altura del footer */
   padding: 20px;
 }
