--- electron-app/src/pages/AdminDashboard.tsx
@@ -1,10 +1,0 @@
-const AdminDashboard = () => {
-  return (
-    <div>
-      <h1>Admin Dashboard</h1>
-      <p>Welcome to the admin dashboard. Here you can manage users, view reports, and configure settings.</p>
-    </div>
-  );
-}
-
-export default AdminDashboard;
