--- electron-app/src/pages/Alumnos.tsx
@@ -1,11 +1,0 @@
-const Alumnos = () => {
-    return (
-        <div className="flex flex-col items-center justify-center h-screen">
-            <h1 className="text-4xl font-bold mb-4">Alumnos</h1>
-            <p className="text-lg">Esta página está en construcción.</p>
-            <p className="text-lg">Pronto podrás ver la lista de alumnos aquí.</p>
-        </div>
-    );
-}
-
-export default Alumnos;
