--- electron-app/src/pages/GastosYProductos.tsx
@@ -1,10 +1,0 @@
-const GastosYProductos = () => {
-  return (
-    <div className="flex flex-col items-center justify-center h-full">
-      <h1 className="text-2xl font-bold mb-4">Gastos y Productos</h1>
-      <p className="text-lg">¡Pronto estará disponible!</p>
-    </div>
-  );
-}
-
-export default GastosYProductos;
