--- electron-app/src/pages/Pacientes.tsx
@@ -1,11 +1,0 @@
-const Pacientes = () => {
-    return (
-        <div className="flex flex-col items-center justify-center h-screen">
-            <h1 className="text-4xl font-bold mb-4">Pacientes</h1>
-            <p className="text-lg">Esta página está en construcción.</p>
-            <p className="text-lg">Pronto podrás gestionar tus pacientes aquí.</p>
-        </div>
-    );
-}
-
-export default Pacientes;
