--- electron-app/src/vite.config.ts
@@ -1,21 +1,0 @@
-import { defineConfig } from 'vite';
-import react from '@vitejs/plugin-react';
-import tailwindcss from '@tailwindcss/vite';
-import { resolve } from 'path';
-
-export default defineConfig({
-  plugins: [react(), tailwindcss()],
-  base: './',
-  build: {
-    outDir: '../dist',
-    emptyOutDir: true,
-  },
-  resolve: {
-    alias: {
-      '@': resolve(__dirname, '.'),
-    },
-  },
-  optimizeDeps: {
-    exclude: ['electron'],
-  },
-});
