+++ electron-app/tsconfig.json
@@ -2,9 +2,9 @@
   "compilerOptions": {
     "jsx": "react-jsx",
     "outDir": "dist",
-    "module": "NodeNext",
-    "target": "ES2022",
-    "moduleResolution": "NodeNext",
+    "module": "ESNext",
+    "target": "ESNext",
+    "moduleResolution": "Node",
     "esModuleInterop": true,
     "strict": true,
     "skipLibCheck": true,
